+++ src/components/Code/Code.story.js
@@ -29,13 +29,21 @@ stories
     <div id="root-preview">
       {story()}
     </div>
-  ));
+  ))
+  .addParameters({
+    knobs: {
+      // Allows users to paste JSX into the text knobs (e.g. `<MyComponent>`)
+      // w/o it getting rewritten as `&lt;MyComponent&gt`;
+      escapeHTML: false,
+    },
+  });
 
 stories
   .add('Default', (() => {
     return (
       <Code
         hasCopyButton={ boolean('hasCopyButton', true) }
+        isHighlighted={ boolean('isHighlighted', true) }
         testSection='my-code-box'
         type={ select('type', {inline: 'inline', block: 'block'}, 'block') }
         language={ select('language', langOptions, 'js') }>
@@ -48,6 +56,7 @@ stories
       <Code
         copyButtonStyle="none"
         hasCopyButton={ boolean('hasCopyButton', true) }
+        isHighlighted={ true }
         testSection='my-code-box'
         type={ select('type', {inline: 'inline', block: 'block'}, 'block') }
         language={ select('language', langOptions, 'js') }>